% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildXmat.R
\name{buildXmat}
\alias{buildXmat}
\title{An auxialiary function to quickly construct the polyomial matrix, 
using Horner's rule}
\usage{
buildXmat(x, nc)
}
\arguments{
\item{x}{The base}

\item{nc}{the number of columns}
}
\value{
A matrix with increasing powers of x in the columns
}
\description{
An auxialiary function to quickly construct the polyomial matrix, 
using Horner's rule
}
