% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateCPM.R
\name{calCPM}
\alias{calCPM}
\title{Calculates counts per millions of reads}
\usage{
calCPM(X, const.mult = 1e+06, ...)
}
\arguments{
\item{X}{an object with class 'SingleCellExperiment', 'data.frame', or 'matrix' 
typically containing gene expression data with genes in rows}

\item{const.mult}{a numerical constant indicatin the desired library size for all samples/cells}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a matrix of CPM
}
\description{
Calculates counts per millions of reads
}
\examples{
 
dat <- make.example.data(n.gene = 10,n.sample = 20, n.group = 1, n.batch = 1)
cpm.dat <- calCPM(dat)
}
