% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateCPM.R
\name{calculateCPM}
\alias{calculateCPM}
\title{Calculates counts per millions of reads, possibly with log-transform}
\usage{
calculateCPM(X, const.mult, prior.count)
}
\arguments{
\item{X}{raw data matrix}

\item{const.mult}{a constant to multiply with}

\item{prior.count}{prior count to be added to the zeroes}
}
\value{
a normalized data matrix
}
\description{
Calculates counts per millions of reads, possibly with log-transform
}
