% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInputValidity.R
\name{checkInputValidity}
\alias{checkInputValidity}
\title{Check for data validity}
\usage{
checkInputValidity(
  s.data,
  group,
  batch,
  group.config,
  batch.config,
  w,
  log.CPM.transform,
  prior.count,
  pDE,
  lib.size.params,
  llStat.thrld,
  result.format
)
}
\arguments{
\item{s.data, group, batch, group.config, batch.config, w, log.CPM.transform, prior.count, pDE, lib.size.params, llStat.thrld, result.format}{see ?SPsimSeq}
}
\value{
Throws errors where neede, otherwise returns invisible
}
\description{
Check for data validity
}
