% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectCandGenes.R
\name{chooseCandGenes}
\alias{chooseCandGenes}
\title{select candidate genes}
\usage{
chooseCandGenes(s.data, X, lfc.thrld = 0, llStat.thrld = 10,
  t.thrld = 2.5, carrier.dist = "normal", max.frac.zero = 1,
  max.frac.zeror.diff = Inf, ...)
}
\arguments{
\item{s.data}{a source data (a SingleCellExperiment object)}

\item{X}{a vector of indicators for group memebership of cells/samples}

\item{lfc.thrld}{a numeric value for the minimum fold change for DE genes}

\item{llStat.thrld}{a numeric value for the minimum squared test statistics from a log-linear model
containing X as a covariate to select DE genes}

\item{t.thrld}{a numeric value for the minimum t statistic for DE genes
('mean.diff'= difference in the mean log CPM, 'rank.sum'= a U-statistic for the log CPM)}

\item{carrier.dist}{a character indicating the type of
carrier density (carrier.dist="normal" or carrier.dist="kernel")}

\item{max.frac.zero}{a numeric value between 0 and 1 indicating the maximum fraction of
zero counts that a DE gene should have}

\item{max.frac.zeror.diff}{a numeric value between 0 and 1 indicating the maximum  absolute
difference in the fraction of zero counts between the groups for DE genes}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a list object contating a set of candidate DE and null genes and additional results
}
\description{
More detailed description
}
\examples{
 # example
}
