% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructDens.R
\name{constructDens}
\alias{constructDens}
\title{Construct the cumulative density}
\usage{
constructDens(densList.ii, exprmt.design, DE.ind.ii)
}
\arguments{
\item{densList.ii}{the estimated density parameters}

\item{exprmt.design}{experiment configuration}

\item{DE.ind.ii}{a boolean, is the gene to be DE?}
}
\value{
The cumulative density
}
\description{
Construct the cumulative density
}
