% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estLibSizeDistr.R
\name{estLibSizeDistr}
\alias{estLibSizeDistr}
\title{Estimate log-normal distribution for the library sizes}
\usage{
estLibSizeDistr(LS, batch)
}
\arguments{
\item{LS}{observed library sizes}

\item{batch}{batches}
}
\value{
Estimated log-normal parameter library sizes
}
\description{
Estimate log-normal distribution for the library sizes
}
