% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configExperiment.R
\name{expriment.config}
\alias{expriment.config}
\title{Configure experiment}
\usage{
expriment.config(batch.config = c(0.25, 0.35, 0.4),
  group.config = c(0.6, 0.4), tot.samples = 150, ...)
}
\arguments{
\item{batch.config}{a numerical vector for the marginal fraction of samples in each batch. 
The number of batches to be simulated is equal to the size of the vector.
All values must sum to 1.}

\item{group.config}{a numerical vector for the marginal fraction of samples in each group.
The number of groups to be simulated is equal to the size of the vector. All values must sum to 1.}

\item{tot.samples}{total number of samples to be simulated.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a list object contating the number of groups and batches to be simukated, 
and the experiment configurartion
}
\description{
Configure experiment
}
\examples{
#---- a design with a total of 10 samples/cells from 1 batch and 1 group
expriment.config(batch.config=1, group.config=1, tot.samples=10)

#---- a design with a total of 20 samples/cells from 1 group and 2 batchs with 
# batch 1 has 15 samples/cells and batch 2 has 5
expriment.config(batch.config=c(15/20, 5/20), group.config=1, tot.samples=20)

#---- a design with a total of 20 samples/cells from 1 batch and 2 groups with 
# group 1 has 10 samples/cells and batch 2 has 10
expriment.config(batch.config=1, group.config=c(0.5, 0.5), tot.samples=20)

#---- a design with a total of 30 samples/cells from 2 groups with group 1 has 15 samples 
# and group 2 has 15, and  three batchs with batch 1,2, and 3 have 5, 10, and 15 samples/cells, 
# respectively.
expriment.config(batch.config=c(5/30, 10/30, 15/30), group.config=c(0.5, 0.5), tot.samples=30)

}
