% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractGlmParams.R
\name{extractGlmParams}
\alias{extractGlmParams}
\title{extract the glm parameters}
\usage{
extractGlmParams(llm)
}
\arguments{
\item{llm}{}
}
\value{
a list of regression parameters and variance covariance matrix
}
\description{
extract the glm parameters
}
