% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMat.R
\docType{methods}
\name{extractMat}
\alias{extractMat}
\alias{extractMat,SingleCellExperiment-method}
\alias{extractMat,matrix-method}
\alias{extractMat,data.frame-method}
\title{A function with S4 dispatching to extract the count matrix}
\usage{
extractMat(Y, ...)

\S4method{extractMat}{SingleCellExperiment}(Y, ...)

\S4method{extractMat}{matrix}(Y, ...)

\S4method{extractMat}{data.frame}(Y, ...)
}
\arguments{
\item{Y}{a matrix, data frame or SingleCellExperiment}

\item{...}{additional arguments, currently ignored}
}
\value{
A data matrix with samples in the columns and genes in the rows
}
\description{
A function with S4 dispatching to extract the count matrix
}
