% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitLLmodel.R
\name{fitLLmodel}
\alias{fitLLmodel}
\title{Fit log linear model for each gene}
\usage{
fitLLmodel(yy, ...)
}
\arguments{
\item{yy}{a list object contating a result from obtCount() function for a single gene}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a list object contating the fitted log linear model
}
\description{
Fit log linear model for each gene
}
\examples{
# 
}
