% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitPoisGlm.R
\name{fitPoisGlm}
\alias{fitPoisGlm}
\title{Fast fit Poisson regression}
\usage{
fitPoisGlm(Ny, x, degree, offset)
}
\arguments{
\item{Ny}{vector of counts}

\item{x}{regressor}

\item{degree}{degree of the polynomial}

\item{offset}{offset}
}
\value{
see glm.fit
}
\description{
Fast fit Poisson regression
}
