% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracZeroLogitModel.R
\name{fracZeroLogitModel}
\alias{fracZeroLogitModel}
\title{Extract data and iterate over batches to estimate zero probability models}
\usage{
fracZeroLogitModel(s.data, batch, cpm.data, n.mean.class, minFracZeroes)
}
\arguments{
\item{s.data, cpm.data}{raw and transformed data}

\item{batch}{the batch vector}

\item{n.mean.class}{see zeroProbModel}

\item{minFracZeroes}{minimum fraction of zeroes before zero-inflation is applied}
}
\value{
a list of binomial regression parameters
}
\description{
Extract data and iterate over batches to estimate zero probability models
}
