% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genCopula.R
\name{genCopula}
\alias{genCopula}
\title{Generate a copula instance}
\usage{
genCopula(corMats, exprmt.design)
}
\arguments{
\item{exprmt.design}{Number of batches, and batch vector}

\item{corMat}{list of correlation matrices}
}
\value{
a list of copula instances
}
\description{
Generate a copula instance
}
