% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genLibSizes.R
\name{genLibSizes}
\alias{genLibSizes}
\title{Generate library sizes from log-normal}
\usage{
genLibSizes(n.batch, n.group, fit.ln, config.mat)
}
\arguments{
\item{n.batch}{}

\item{n.group}{}

\item{fit.ln}{}

\item{config.mat}{}
}
\value{
The generated libray sizes per batch and group
}
\description{
Generate library sizes from log-normal
}
