% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeExampleRNAseqData.R
\name{make.example.data}
\alias{make.example.data}
\title{Make an example RNA-seq data.}
\usage{
make.example.data(n.gene = 2000, n.sample = 30, n.group = 2,
  n.batch = 3, p.DEgenes = 0.2, common.BCV = 0.25, type = "bulk")
}
\arguments{
\item{n.gene}{number of genes.}

\item{n.sample}{number of samples/cells.}

\item{n.group}{number of biological group.}

\item{n.batch}{number of batchs.}

\item{p.DEgenes}{fraction of genes with DE between simulated groups}

\item{common.BCV}{common biological coefficient of variation}

\item{type}{type gene expression data, either bulk RNA-seq (type="bulk") or 
single cell RNA-seq (type="sc")}
}
\value{
a SingleCellExperiment object.
}
\description{
Make an example RNA-seq data.
}
\examples{
make.example.data()
 
}
