% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchCopula.R
\name{matchCopula}
\alias{matchCopula}
\title{Match copulas to estimated SP distribution}
\usage{
matchCopula(cumDens, exprmt.design, copSam, sel.genes.ii)
}
\arguments{
\item{cumDens}{The cumulative densities evaluated}

\item{exprmt.design}{the design}

\item{copSam}{the sampled copula}

\item{sel.genes.ii}{the gene}
}
\value{
the outcome values as a vector
}
\description{
Match copulas to estimated SP distribution
}
