% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtCount.R
\name{obtCount}
\alias{obtCount}
\title{Calculates height and mid points of a distribution}
\usage{
obtCount(Y, w)
}
\arguments{
\item{Y}{a vector of gene expression data for a particular gene (in log CPM)}

\item{w}{a numeric value between 0 and 1 or NULL refering the number of classes to be created}
}
\value{
a list object contating class breaks, mid points and counts
}
\description{
Calculates height and mid points of a distribution
}
