% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneParmEst.R
\name{parmEstDensVec}
\alias{parmEstDensVec}
\title{Density estimation on a single vector}
\usage{
parmEstDensVec(
  Y0,
  model.zero.prob,
  min.val,
  w,
  prev.min.val = 0.25,
  min.count.nonnull = 3
)
}
\arguments{
\item{Y0}{the vector of observations}

\item{model.zero.prob, min.val, w}{see geneParmEst()}

\item{prev.min.val}{minimum prevalence of minimum values}

\item{min.count.nonnull}{minimum count for estimation}
}
\value{
density estimates
}
\description{
Density estimation on a single vector
}
