% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareSPsimOutputs.R
\name{prepareSPsimOutputs}
\alias{prepareSPsimOutputs}
\title{A function to prepare outputs}
\usage{
prepareSPsimOutputs(sim.dat, exprmt.design, DE.ind, result.format, LL)
}
\arguments{
\item{sim.dat}{The simulated data}

\item{exprmt.design}{the design}

\item{DE.ind}{the differential abundance indicator}

\item{result.format}{the desired output format}

\item{LL}{simulated library sizes}
}
\value{
the data in the desired format
}
\description{
A function to prepare outputs
}
