% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectGenes.R
\name{selectGenes}
\alias{selectGenes}
\title{Sample genes from candidate genes}
\usage{
selectGenes(pDE, exprmt.design, n.genes, null.genes0, nonnull.genes0)
}
\arguments{
\item{pDE}{fraction of genes to be made DE}

\item{exprmt.design}{the experiment design}

\item{n.genes}{the total number of genes required}

\item{null.genes0, nonnull.genes0}{Candidate null and non-null genes}
}
\value{
a vector of selected genes
}
\description{
Sample genes from candidate genes
}
