% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLibSizes.R
\name{simLibSize}
\alias{simLibSize}
\title{A function simulate library sizes}
\usage{
simLibSize(sub.batchs, LS, lib.size.params, n.batch, batch, n.group,
  config.mat)
}
\arguments{
\item{sub.batchs}{}

\item{LS}{observed library sizes}

\item{lib.size.params}{Parameters of the lognormal}

\item{n.batch}{number of batches}

\item{batch}{batches}

\item{n.group}{number of groups}

\item{config.mat}{groups}
}
\value{
simulated library sizes
}
\description{
A function simulate library sizes
}
