% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{simulateData}
\alias{simulateData}
\title{A function to simulate bulk or single cell RNA sequencing data}
\usage{
simulateData(n.sim = 1, s.data, batch = NULL, group = NULL,
  n.genes = 1000, batch.config = 1, group.config = c(0.5, 0.5),
  tot.samples = 150, pDE = 0.2, model.zero.prob = FALSE,
  result.format = "SCE", verbose = TRUE, ...)
}
\arguments{
\item{n.sim}{a numerical value for the number of simulated data to be  generated}

\item{s.data}{a source data (a SingleCellExperiment class object or a matrix/data.frame of counts with genes in 
rows and samples in columns)}

\item{batch}{a vector containg btach indicator for each sample/cell}

\item{group}{a vector containg group indicator for each sample/cell}

\item{n.genes}{a numeric value for the total number of genes to be simulated}

\item{batch.config}{a numerical vector for the marginal fraction of samples in each batch. 
The number of batches to be simulated is equal to the size of the vector.
All values must sum to 1.}

\item{group.config}{a numerical vector for the marginal fraction of samples in each group.
The number of groups to be simulated is equal to the size of the vector. All values must sum to 1.}

\item{tot.samples}{a numerical value for total number of samples to be simulated.}

\item{pDE}{a numeric value between 0 and 1 indicating the fraction of DE genes 
in a single simulated data}

\item{model.zero.prob}{a logical value whether to model the zero probablity separately 
(suitable for single cell data)}

\item{result.format}{a character value for the type of format for the output. Choice can  be
'SCE' for SingleCellExperiment class or "list" for a list object that contains the simulated count,
column information abd row information.}

\item{verbose}{a logical value, if TRUE it displays a message about the satatus of the simulation}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a list of SingleCellExperiment object each contatining simulated counts (not normalized), 
cell level information in colData, and gene level information in rowData.
}
\description{
This function simulates RNA sequencing data given a real RNA-seq data using
semi-parametric density estimation.
}
\details{
This function constructs the actual density of a given bulk or single cell RNA-seq data 
(passed using \emph{s.data} argument) using a specially designed exponetial family for density
estimation. Afterwards, a new data set is simulated from the estimated density. The objective is 
to maximaly retainthe characterisics of the source data in the simulated data without distributional 
assumption to the gene expression. This is particulaly useful for cases with complex/unknown 
distributions of the data, such as for single cell RNA-seq.

For single cell RNA-seq data, the high abundance of zero counts is a special characteristic as
a result of either biological or technical factors. Therefore, this package has a special feature to
model the probability of zero counts given the mean expression and library size  to add excess zeros
representing the technical effectss. This can be achieved by using \emph{model.zero.prob=TRUE}. Note that,
for extremly large size data, this feature may not work and a rrandom sample of cells (up to 400) can
be used to fit the model. To enable sub-set the data, add the argument \emph{subset.data=TRUE} and you 
can specify the number of cells to be used using \emph{n.samples} argument. For example \emph{n.samples=300}.

Different experimental designs can be simulated using the group and batch configuration arguments to
simulate biologica/experimental conditions and batchs (instrument or subjects), respectively. However,
the design of the new simulated data strongly depends on the design in the source data. In addition,
batch simulation may lead to unexpected error if the number of batches in the source data is small (e.g.
less than 5) and/or there is no enough data in each batch.
}
\examples{
##
}
\references{
\itemize{
\item Efron, B., & Tibshirani, R. (1996). Using specially designed exponential families for density estimation. \emph{The Annals of Statistics}, 24(6), 2431-2461.
}
}
