% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPsimPerGene.R
\name{SPsimPerGene}
\alias{SPsimPerGene}
\title{A function that generates the simulated data for each gene}
\usage{
SPsimPerGene(cpm.data, est.list.ii, DE.ind.ii, n.batch, n.group, batch,
  group, config.mat, sel.genes.ii, log.CPM.transform, const, min.val,
  null.group, LL, copulas.batch, tot.samples, model.zero.prob,
  fracZero.logit.list)
}
\arguments{
\item{cpm.data}{}

\item{est.list.ii}{}

\item{DE.ind.ii}{}

\item{n.batch}{}

\item{n.group}{}

\item{batch}{}

\item{group}{}

\item{config.mat}{}

\item{sel.genes.ii}{}

\item{log.CPM.transform}{}

\item{const}{}

\item{min.val}{}

\item{null.group}{}

\item{LL}{}

\item{copulas.batch}{}

\item{tot.samples}{}

\item{model.zero.prob}{}

\item{fracZero.logit.list}{}
}
\description{
A function that generates the simulated data for each gene
}
