% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPsimSeq-package.R
\docType{package}
\name{SPsimSeq-package}
\alias{SPsimSeq-package}
\title{SPsimSeq package}
\description{
Using a specially designed exponential family for density estimation 
it constructs the distribution of gene expression from a given real RNA 
sequencing data (single and bulk). Subsequently, it simulates a new 
dataset from the estimated density. It allows simulation of multiple 
groups and batches with any required sample size and library size.
}
\author{
Alemu Takele Assefa \email{alemutak@hotmail.com}
}
