% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{SPsimSeqData}
\alias{SPsimSeqData}
\title{A function to simulate bulk or single cell RNA sequencing data}
\usage{
SPsimSeqData(n.sim = 1, s.data, batch = NULL, group = NULL,
  cand.genes = NULL, n.genes = 1000, batch.config = c(0.33, 0.33,
  0.34), group.config = c(0.5, 0.5), tot.samples = 150, pDE = 0.2,
  model.zero.prob = FALSE, null.group = which.max(table(group)),
  fc.type = "mean.diff", lfc.thrld = 0.5, U.thrld = 0.7,
  llStat.thrld = 5, t.thrld = 2.5, carrier.dist = "normal",
  w = 0.5, max.frac.zero = 0.7, max.frac.zeror.diff = 0.1,
  n.mean.class = 30, subset.data = FALSE, n.samples = 400,
  verbose = TRUE, ...)
}
\arguments{
\item{n.sim}{number of simulated data to be created}

\item{s.data}{a source data (a SingleCellExperiment object)}

\item{batch}{a vector containg btach indicator for each sample/cell}

\item{group}{a vector containg group indicator for each sample/cell}

\item{cand.genes}{a list object contating candidate genes}

\item{n.genes}{a numeric value for the total number of genes to be simulated}

\item{batch.config}{a numerical vector for the marginal fraction of samples in each batch. 
The number of batches to be simulated is equal to the size of the vector.
All values must sum to 1.}

\item{group.config}{a numerical vector for the marginal fraction of samples in each group.
The number of groups to be simulated is equal to the size of the vector. All values must sum to 1.}

\item{tot.samples}{total number of samples to be simulated.}

\item{pDE}{a numeric value between 0 and 1 indicating the fraction of DE genes 
in a single simulated data}

\item{model.zero.prob}{a logical value whether to model the zero probablity separately 
(suitable for single cell data)}

\item{null.group}{a character value that indicating a name of}

\item{fc.type}{a character indicating how DE genes should be calculated 
('mean.diff'= difference in the mean log CPM, 'rank.sum'= a U-statistic for the log CPM)}

\item{lfc.thrld}{a numeric value for the minimum fold change for DE genes (if fc.type='mean.diff')}

\item{U.thrld}{a numeric value for the minimum U-statistic for DE genes (if fc.type='rank.sum')}

\item{llStat.thrld}{a numeric value for the minimum squared test statistics from a log-linear 
model containing X as a covariate to select DE genes}

\item{t.thrld}{a numeric value for the minimum t statistic for DE genes (if fc.type='mean.diff')}

\item{carrier.dist}{a character indicating the type of carrier density (carrier.dist="normal" 
or carrier.dist="kernel")}

\item{w}{a numeric value between 0 and 1 or NULL refering the number of classes to be created 
for the outcome data (if NULL the algorithm to calculate breakes in graphics::hist() function 
will be used)}

\item{max.frac.zero}{a numeric value between 0 and 1 indicating the maximum fraction of zero 
counts that a DE gene should have}

\item{max.frac.zeror.diff}{a numeric value between 0 and 1 indicating the maximum  absolute 
difference in the fraction of zero counts between the groups for DE genes}

\item{n.mean.class}{an integer for the number of groups to be created for the mean log CPM of genes}

\item{subset.data}{a logical value to subset columns(samples) if the size of the data is too big
for space and computation time saving}

\item{n.samples}{an integer indicating the number of samples/cells to sample (if subset.data=TRUE)}

\item{verbose}{a logical value, if TRUE it displays a message about the satatus of the simulation}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a list of SingleCellExperiment object  each contatining simulated counts (not normalized), 
cell level information in colData, and gene level information in rowData
}
\description{
A function to simulate bulk or single cell RNA sequencing data
}
\examples{
##
}
