% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroProbModel.R
\name{zeroProbModel}
\alias{zeroProbModel}
\title{predict zero probability using logistic rgression}
\usage{
zeroProbModel(count.data, n.mean.class = 30, subset.data = FALSE,
  n.samples = 400, ...)
}
\arguments{
\item{count.data}{a count data matrix with rows represent genes and columns represent cells/samples}

\item{n.mean.class}{an integer for the number of groups to be created for the mean log CPM of genes}

\item{subset.data}{a logical value to subset columns(samples) if the size of the data is too big
for space and computation time saving}

\item{n.samples}{an integer indicating the number of samples/cells to sample (if subset.data=TRUE)}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a GLM class object contating the estimated logistic regression
}
\description{
More detailed description
}
\examples{
 # example
}
