% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectCandGenes.R
\name{chooseCandGenes}
\alias{chooseCandGenes}
\title{Select candidate genes}
\usage{
chooseCandGenes(cpm.data, X, lfc.thrld = 1, llStat.thrld = 10,
  t.thrld = 2.5, max.frac.zeror.diff = Inf, const = 1e-05, ...)
}
\arguments{
\item{cpm.data}{CPM matrix}

\item{X}{a vector of indicators for group memebership of cells/samples}

\item{lfc.thrld}{a numeric value for the minimum fold change for DE genes}

\item{llStat.thrld}{a numeric value for the minimum squared test statistics from a log-linear model
containing X as a covariate to select DE genes}

\item{t.thrld}{a numeric value for the minimum t statistic for DE genes}

\item{max.frac.zeror.diff}{a numeric value >=0 indicating the maximum  absolute
difference in the fraction of zero counts between the groups for DE genes. Default in Inf}

\item{const}{a small constant (>0) added to the CPM before log transformation, to avoid  log(0).
default is 1e-5, see in}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a list object contating a set of candidate null and non-null genes and additional results
}
\description{
This function can be used independently to select candidate genes from a given real RNA-srq data (bulk/single)
for the SPsimSeq simulation. It chooses genes with cetrail chracteristics, such as log-fold-change 
above a certain thereshold.
}
\examples{
 # example
}
