% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitLLmodel.R
\name{fitLLmodel}
\alias{fitLLmodel}
\title{Fit log linear model for each gene}
\usage{
fitLLmodel(yy, mu.hat, sig.hat, n)
}
\arguments{
\item{yy}{a list object contating a result from obtCount() function for a single gene}

\item{mu.hat, sig.hat}{Carrier density estimators}

\item{n}{number of observations}
}
\value{
a list object containing the fitted log linear model and carrier density
}
\description{
Fit log linear model for each gene
}
