% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genCopula.R
\name{genCopula}
\alias{genCopula}
\title{Generate a copula instance}
\usage{
genCopula(corMat, n.batch, batch)
}
\arguments{
\item{corMat}{list of correlation matrices}

\item{n.batch, batch}{Number of batches, and batch vector}
}
\value{
a list of copula instances
}
\description{
Generate a copula instance
}
