% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneParmEst.R
\name{geneParmEst}
\alias{geneParmEst}
\title{Gene level param estimates for density estimation}
\usage{
geneParmEst(
  cpm.data.i,
  batch,
  group,
  prior.count = prior.count,
  de.ind,
  model.zero.prob,
  w
)
}
\arguments{
\item{cpm.data.i}{full vector of genewise observation}

\item{batch, group}{batch and grouping vectors}

\item{prior.count}{the prior count for the cpm transofrm}

\item{de.ind}{a boolean, is the gene to be DE?}

\item{model.zero.prob}{a boolean, should zero-density be modelled?}

\item{w}{weight}
}
\value{
list of density estimates
}
\description{
Gene level param estimates for density estimation
}
