% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genLibSizes.R
\name{genLibSizes}
\alias{genLibSizes}
\title{Generate library sizes from log-normal}
\usage{
genLibSizes(fit.ln, exprmt.design)
}
\arguments{
\item{fit.ln}{the library size model}

\item{exprmt.design}{the design}
}
\value{
The generated libray sizes per batch and group
}
\description{
Generate library sizes from log-normal
}
