% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchCopula.R
\name{matchCopula}
\alias{matchCopula}
\title{Match copulas to estimated SP distribution}
\usage{
matchCopula(cumDens, exprmt.design, copSam, sel.genes.ii)
}
\arguments{
\item{cumDens}{}

\item{exprmt.design}{}

\item{copSam}{}

\item{sel.genes.ii}{}
}
\value{
the outcome values as a vector
}
\description{
Match copulas to estimated SP distribution
}
