% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtCorMatsBatch.R
\name{obtCorMatsBatch}
\alias{obtCorMatsBatch}
\title{A function to obtain copulas or uniform random variables}
\usage{
obtCorMatsBatch(cpm.data, batch)
}
\arguments{
\item{cpm.data}{the transformed data matrix}

\item{batch}{the batch indicators}
}
\value{
The estimated correlation matrices per batch
}
\description{
A function to obtain copulas or uniform random variables
}
