% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtCount.R
\name{obtCount}
\alias{obtCount}
\title{Calculates height and mid points of a distribution}
\usage{
obtCount(Y, w = NULL, ...)
}
\arguments{
\item{Y}{a vector of gene expression data for a particular gene (in log CPM)}

\item{w}{a numeric value between 0 and 1 or NULL refering the number of classes to be created
for the outcome data (if NULL the algorithm in graphics::hist() function will be used)}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a list object contating class mid points, counts, and others
}
\description{
Calculates height and mid points of a distribution
}
\examples{
#
}
