% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareSourceData.R
\name{prepareSourceData}
\alias{prepareSourceData}
\title{Prepare source data for simulation.}
\usage{
prepareSourceData(s.data, batch = NULL, group = NULL,
  cand.genes = NULL, exprmt.design, simCtr, ...)
}
\arguments{
\item{s.data}{a source data (a SingleCellExperiment object)}

\item{batch}{a vector containg btach indicator for each sample/cell}

\item{group}{a vector containg group indicator for each sample/cell}

\item{cand.genes}{a list object contating candidate genes}

\item{exprmt.design}{a list contating simulation design configuration}

\item{simCtr}{(integer) seed number}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a list object
}
\description{
Prepare source data for simulation.
}
\examples{
 # example 
}
