% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareSourceData.R
\name{prepareSourceData}
\alias{prepareSourceData}
\title{Prepare source data for simulation.}
\usage{
prepareSourceData(s.data, batch, group, cand.DE.genes, exprmt.design,
  const, lfc.thrld, llStat.thrld, t.thrld, w, log.CPM.transform, ...)
}
\arguments{
\item{s.data}{a source data matrix}

\item{batch}{a vector containg btach indicator for each sample/cell}

\item{group}{a vector containg group indicator for each sample/cell}

\item{cand.DE.genes}{a list object contating candidate predictor (DE) genes}

\item{exprmt.design}{a list contating simulation design configuration}

\item{const}{a small constant (>0) to be added to the CPM before log transformation, to avoid  log(0).
default is 1e-5}

\item{lfc.thrld, llStat.thrld, t.thrld, w}{see ?chooseCandGenes}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a list object
}
\description{
Prepare source data for simulation.
}
\examples{
 # example 
}
