% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGPscData.R
\docType{data}
\name{scNGP.data}
\alias{scNGP.data}
\title{Neuroblastoma cell line single cell RNA-seq.}
\format{A SingleCellExperiment object}
\source{
\url{http://dx.doi.org/10.1101/430090}
}
\usage{
scNGP.data
}
\description{
Contains 83 cell line neuroblastoma  cells (31 nutlin-3 treated and 52 controls). The data 
is generated using SMARTer/C1 protocol.
}
\references{
Verboom, K., Everaert, C., Bolduc, N., Livak, K. J., Yigit, N., Rombaut, D., ... & Mestdagh, P. (2018). SMARTer single cell total RNA sequencing. BioRxiv, 430090.
\describe{
  \item{SingleCellExperiment}{counts + gene info+cell infro} 
}
}
\keyword{datasets}
