% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPsimPerGene.R
\name{SPsimPerGene}
\alias{SPsimPerGene}
\title{A function that generates the simulated data for each gene}
\usage{
SPsimPerGene(cumDens, exprmt.design, sel.genes.ii, log.CPM.transform,
  prior.count, LL, copSam, model.zero.prob, fracZero.logit.list,
  const.mult)
}
\arguments{
\item{cumDens}{}

\item{sel.genes.ii}{}

\item{log.CPM.transform}{}

\item{LL}{}

\item{copSam}{}

\item{model.zero.prob}{}

\item{fracZero.logit.list}{}
}
\value{
Simulated cpm values
}
\description{
A function that generates the simulated data for each gene
}
