% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPsim.R
\docType{package}
\name{SPsimSeq}
\alias{SPsimSeq}
\alias{SPsimSeq-package}
\title{SPsim package}
\description{
It uses specially designed exponential family for density estimation
from which RNA sequencing data will be simulated. The first step of is
estimting the actual density of a real sequencing data using log-linear
model and new data set is sampled from the estimated density.
}
\author{
Alemu Takele Assefa \email{alemutak@hotmail.com}
}
