% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeroProbModel.R
\name{zeroProbModel}
\alias{zeroProbModel}
\title{Predict zero probability using logistic rgression}
\usage{
zeroProbModel(cpm.data, logL, zeroMat, n.mean.class)
}
\arguments{
\item{cpm.data}{log CPM matrix}

\item{logL}{log library size of the source data}

\item{zeroMat}{the matrix of zero indicators}

\item{n.mean.class}{a fraction of the number of genes
for the number of groups to be created for the mean log CPM of genes}
}
\value{
The coefficients of the estimated logistic regression
}
\description{
Predict zero probability using logistic rgression
}
